move ValueCollection to engdb_lib
Clean out some cruft
